# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CortexDB Beta API
  version: 0.2.0-beta.2
  description: 'Beta HTTP API for the local single-node CortexDB
    engine. The HTTP layer is async, while the database core is a local blocking engine
    executed through a blocking task boundary.

    '
servers:
- url: http://127.0.0.1:8181
paths:
...
  /v1/cluster/status:
    get:
      summary: Node and replication status
      description: Returns cluster layout currently known to this node.
      responses:
        '200':
          description: Cluster status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterStatusResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
...
  schemas:
...
    ClusterNodeResponse:
      type: object
      required:
      - id
      - address
      properties:
        id:
          type: integer
          format: uint64
        address:
          type: string
    ClusterStatusResponse:
      type: object
      required:
      - local_node
      - nodes
      - replication_factor
      - distributed_enabled
      properties:
        local_node:
          type: integer
          format: uint64
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/ClusterNodeResponse'
        replication_factor:
          type: integer
        distributed_enabled:
          type: boolean
...
    ErrorResponse:
      type: object
      required:
      - code
      - error
      - message
      properties:
        code:
          type: string
          enum:
          - not_found
          - bad_request
          - unauthorized
          - forbidden
          - payload_too_large
          - rate_limited
          - quota_exceeded
          - service_unavailable
          - internal
          - invalid_aql
          - unknown_field
          - unsupported_operator
          - permission_denied
          - database_busy
          - storage_corruption
          - invalid_tenant
        error:
          type: string
        message:
          type: string
...
  responses:
...
    InternalError:
      description: Internal Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'